feat: add readme

+++ app/models/leads_model.py
@@ -4,7 +4,24 @@ from datetime import date, datetime
 from app.configs.database import db
 from flask import current_app
 from sqlalchemy import Column, Date, String
-from sqlalchemy.sql.sqltypes import Integer
+from sqlalchemy.sql.sqltypes import DateTime, Integer
+
+
+# - trabalhar com as seguintes EXCEÇÕES da API:
+# - campo não existe, campo faltando, dado de um campo em formato inválido
+# - EXCEÇÕES DO SQL:
+
+#     -usuário não existe
+
+#     EXCEÇÕES DE REGRAS DE NEGÓCIO
+
+#     -ex: pediram data em formato específico, ex 22-01-2021 22:12:04
+
+#     -não permitir data do futuro
+# pode pegar a mensagem do TypeError as e e return str(e)
+
+# verificar cada um dos métodos pra ver se eles precisam mesmo existir
+# se não rola de chamar direto pelo flask_SQLAlchemy e ORM que nem o Lucas fez na demo
 
 @dataclass
 class Leads(db.Model):
@@ -20,8 +37,8 @@ class Leads(db.Model):
     name = Column(String, nullable=False)
     email = Column(String, nullable=False, unique=True )
     phone = Column(String, unique=True, nullable=False)
-    creation_date = Column(Date(), nullable=False)
-    last_visit = Column(Date(), nullable=True)
+    creation_date = Column(DateTime, nullable=False)
+    last_visit = Column(DateTime, nullable=True)
     visits = Column(Integer, nullable=True, default=1)
     
     
@@ -32,6 +49,7 @@ class Leads(db.Model):
         self.creation_date = datetime.utcnow()
         self.last_visit = datetime.utcnow()
         self.visits = 1
+        
         
     def query_to_json(self):
         return {"name": self.name, "email":self.email, "phone": self.phone, "creation_date": self.creation_date, "last_visit": self.last_visit, "visits": self.visits} 
@@ -62,11 +80,12 @@ class Leads(db.Model):
 
     @staticmethod    
     def get_all_users():
+
       data = Leads.query.all()
       data_to_json = [ vaccine.query_to_json() for vaccine in data]
       sorted_data =  sorted(data_to_json, key=lambda k: k['visits'], reverse=True) 
       return sorted_data
-
+            
     @staticmethod
     def check_data_key(**kwargs):
         data_keys = [i for i in kwargs.keys()]
@@ -79,14 +98,18 @@ class Leads(db.Model):
         user.last_visit = datetime.utcnow()
         user.visits += 1
         db.session.commit()
-        return user
+        data = user.query_to_json()
+        return data
 
     @staticmethod    
     def delete_user_data(**kwargs):
-        user = Leads.query.filter_by(email=kwargs["email"]).first()        
+        user = Leads.query.filter_by(email=kwargs["email"]).first()  
+        if user == None:
+          return False      
         db.session.delete(user)
         db.session.commit()
-        return user
+        data = user.query_to_json()
+        return data
